changes to fix DeleteRecordIfFieldEmpty behavior when id is something other than   id = 001, first
SolrMarcIndexerExceptions no longer will terminate the indexing of a given record, the entire index map will be computed and a special case value will be added to indicate the Exception condition.   Then just before the index map is added to Solr these special values will be looked-for and the correct action will be taken.

git-svn-id: http://solrmarc.googlecode.com/svn/trunk@1642 31eaa32e-b84e-0410-9ca4-7366b6a8ecef